update flow layout view controller
