--- InternTuan/InternTuan/Issues/Issue 13 (Collection view)/issue 13 (flow layout)/view/cell/Issue13FlowLayoutViewController.xib
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="24127" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="24063"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="Issue13FlowLayoutViewController" customModule="InternTuan" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="Fap-Uk-vQp" id="E9R-LT-Bdw"/>
-                <outlet property="hayLabel" destination="7iF-iv-lUf" id="Cy6-lu-mvo"/>
-                <outlet property="selectModeButton" destination="3Is-a1-v6a" id="tc0-UE-Bbw"/>
-                <outlet property="view" destination="iN0-l3-epB" id="PPF-GU-vUc"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="Fap-Uk-vQp">
-                    <rect key="frame" x="16" y="179" width="361" height="605"/>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    <collectionViewFlowLayout key="collectionViewLayout" automaticEstimatedItemSize="YES" minimumLineSpacing="10" minimumInteritemSpacing="10" id="AVb-qD-Yba">
-                        <size key="itemSize" width="128" height="128"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Hay" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7iF-iv-lUf">
-                    <rect key="frame" x="150" y="138" width="93" height="21"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="3Is-a1-v6a">
-                    <rect key="frame" x="307" y="129" width="70" height="39"/>
-                    <state key="normal" title="Button"/>
-                    <buttonConfiguration key="configuration" style="filled" title="select"/>
-                    <connections>
-                        <action selector="tapSelect" destination="-1" eventType="touchUpInside" id="vRA-CB-u7I"/>
-                    </connections>
-                </button>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="7iF-iv-lUf" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="20" id="6Jf-U2-qWR"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="7iF-iv-lUf" secondAttribute="trailing" constant="150" id="A3S-au-nt7"/>
-                <constraint firstItem="3Is-a1-v6a" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="11" id="FnG-yL-ryW"/>
-                <constraint firstItem="Fap-Uk-vQp" firstAttribute="top" secondItem="7iF-iv-lUf" secondAttribute="bottom" constant="20" id="Qe6-dI-PDd"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="3Is-a1-v6a" secondAttribute="trailing" constant="16" id="QjL-Be-Vda"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="Fap-Uk-vQp" secondAttribute="bottom" id="T0J-AV-nLY"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="Fap-Uk-vQp" secondAttribute="trailing" constant="16" id="j6X-bs-HUO"/>
-                <constraint firstItem="7iF-iv-lUf" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="150" id="lLo-oK-Y9b"/>
-                <constraint firstItem="Fap-Uk-vQp" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="ypV-Sw-nt6"/>
-            </constraints>
-            <point key="canvasLocation" x="8" y="125"/>
-        </view>
-    </objects>
-    <resources>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-    </resources>
-</document>
